testing with timeout and flating ip cleanup
Why:
It seems nodepool is trying to reach the instance after the floating IP was removed from the instance. Therefore, we start seeing paramiko error (no valid host key) probably because it can no longer reach the systems.

I will try with the option to not clean the flaoting IP and try to reduce the timeout.